Move KB class and func. to own file
